Code Modular:Code shifter to common Service
